fix: add support of RawAudio in FeatureExtractionWrapper (was Pre___ed)
