+++ pyannote/audio/pipeline/overlap_detection.py
@@ -48,7 +48,7 @@ from pyannote.metrics.detection import DetectionPrecision
 from pyannote.metrics.detection import DetectionRecall
 from pyannote.metrics.detection import DetectionPrecisionRecallFMeasure
 from pyannote.metrics import f_measure
-from pyannote.audio.utils.path import Pre___ed
+from pyannote.audio.features.wrapper import FeatureExtractionWrapper
 
 
 class OverlapDetection(Pipeline):
@@ -87,7 +87,7 @@ class OverlapDetection(Pipeline):
         if scores is None:
             scores = "@ovl_scores"
         self.scores = scores
-        self._scores = Pre___ed(self.scores)
+        self._scores = FeatureExtractionWrapper(self.scores)
 
         self.precision = precision
         self.fscore = fscore
