+++ pyannote/audio/pipeline/speaker_change_detection.py
@@ -46,7 +46,7 @@ from pyannote.database import get_unique_identifier
 from pyannote.metrics.segmentation import SegmentationPurityCoverageFMeasure
 from pyannote.metrics.diarization import DiarizationPurityCoverageFMeasure
 
-from pyannote.audio.utils.path import Pre___ed
+from pyannote.audio.features.wrapper import FeatureExtractionWrapper
 
 
 class SpeakerChangeDetection(Pipeline):
@@ -86,7 +86,7 @@ class SpeakerChangeDetection(Pipeline):
         if scores is None:
             scores = "@scd_scores"
         self.scores = scores
-        self._scores = Pre___ed(self.scores)
+        self._scores = FeatureExtractionWrapper(self.scores)
 
         self.purity = purity
         self.fscore = fscore
